Unit tests and debugging
Started writing unit tests, separated some classes into extra files and
did a bit of debugging.
