+++ WPFSKillTree/WPFSKillTree.csproj
@@ -161,7 +161,9 @@
     <Compile Include="SkillTreeFiles\SkillIcons.cs" />
     <Compile Include="SkillTreeFiles\SkillNode.cs" />
     <Compile Include="SkillTreeFiles\SkillNodeGroup.cs" />
+    <Compile Include="SkillTreeFiles\SteinerTrees\DistanceLookup.cs" />
     <Compile Include="SkillTreeFiles\SteinerTrees\GeneticAlgorithm.cs" />
+    <Compile Include="SkillTreeFiles\SteinerTrees\Graphs.cs" />
     <Compile Include="SkillTreeFiles\SteinerTrees\HeapPriorityQueue.cs" />
     <Compile Include="SkillTreeFiles\SteinerTrees\IPriorityQueue.cs" />
     <Compile Include="SkillTreeFiles\SteinerTrees\PriorityQueueNode.cs" />
